+++ open_alchemy/table_args/factory.py
@@ -72,13 +72,13 @@ with open(_SCHEMAS_FILE) as in_file:
 _UNIQUE_SCHEMA_NAMES: typing.List[str] = list(
     map(
         lambda schema: schema["$ref"].split("/")[-1],
-        _SCHEMAS["x-composite-unique"]["oneOf"],
+        _COMMON_SCHEMAS["CompositeUnique"]["oneOf"],
     )
 )
 _INDEX_SCHEMA_NAMES: typing.List[str] = list(
     map(
         lambda schema: schema["$ref"].split("/")[-1],
-        _SCHEMAS["x-composite-index"]["oneOf"],
+        _COMMON_SCHEMAS["CompositeIndex"]["oneOf"],
     )
 )