+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Add optional parameter that can be used to generate a models file for IDE auto complete and type hinting
 - Add _from_dict_ and _to_dict_ to the type models file
 - Add SQLAlchemy information to models file
+- Add back references to models file
 
 ## Version 0.10.4 - 2019-12-18